pyproject.toml
[project]
name = "stryx"
version = "0.1.0"
description = "Typed configuration compiler for ML experiments"
readme = "README.md"
authors = [
    { name = "Filippos Christianos", email = "filippos.christianos@gmail.com" }
]
license = { text = "MIT" }
repository = "https://github.com/fchristianos/stryx"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.11"
dependencies = [
    "filelock>=3.0",
    "petname>=2.6",
    "prompt-toolkit>=3.0.52",
    "pydantic>=2",
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.0.0",
]

[project.scripts]
stryx = "stryx.cli:main"

[project.optional-dependencies]
examples = [
    "trl>=0.26.1",
]
dev = [
    "pytest>=8.0",
    "torch>=2.0.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
]

[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
